requirements.txt
python==3.9.18
streamlit==1.13.0
google-generativeai==0.3.2